+++ store/postgres/postgres.go
@@ -81,6 +81,7 @@ var _ oasis.KeywordSearcher = (*Store)(nil)
 var _ oasis.GraphStore = (*Store)(nil)
 var _ oasis.BidirectionalGraphStore = (*Store)(nil)
 var _ oasis.CheckpointStore = (*Store)(nil)
+var _ oasis.DocumentMetaLister = (*Store)(nil)
 
 // nopLogger is a logger that discards all output.
 var nopLogger = slog.New(pgDiscardHandler{})